Refactor Logic
create a list of buttons using querySelectorAll

Loop through button (NodeList?) to add identical event listener to each.

Add/Remove “selected” class to show highlighting

Create a “reset” function to be called in in both the button event
listener and the reset button that refreshes the colors available for
choosing as well as the appropriate number of squares.